Refactor product components
